python/requirements.txt
python-fasthtml
uvicorn
mcp
playwright
pywin32; sys_platform == "win32"
